+++ chat.js
@@ -76,6 +76,8 @@ class Chat extends BaseClaude {
 		this._cumulativeUsage = {
 			promptTokens: this.lastResponseMetadata.promptTokens,
 			responseTokens: this.lastResponseMetadata.responseTokens,
+			cacheCreationTokens: this.lastResponseMetadata.cacheCreationTokens,
+			cacheReadTokens: this.lastResponseMetadata.cacheReadTokens,
 			totalTokens: this.lastResponseMetadata.totalTokens,
 			attempts: 1
 		};